Put a guard around os.remove call on sorted input .bam

+++ nimble/__main__.py
@@ -158,8 +158,9 @@ def align(param_list):
         sys.stdout.flush()
         subprocess.call([path] + param_list)
 
-        print("Deleting intermediate sorted .bam file")
-        os.remove(param_list[bam_param_idx])
+        if input_ext == ".bam":
+            print("Deleting intermediate sorted .bam file")
+            os.remove(param_list[bam_param_idx])
     else:
         print("No aligner found. Attempting to download the latest release.\n")
         download([])
